coffeeAGNTCY/coffee_agents/lungo/pyproject.toml
[project]
name = "lungo"
version = "0.1.0"
description = ""
authors = [{ name = "Shridhar Shah", email = "shridhsh@cisco.com" }]
requires-python = ">=3.12,<4.0"
dependencies = [
    "a2a-sdk==0.3.0",
    "click>=8.1.8",
    "coloredlogs>=15.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain-anthropic>=0.3.13",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "langgraph-supervisor>=0.0.26",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests",
    "starlette",
    "uvicorn",
    "mcp[cli]>=1.10.0",
    "ioa-observe-sdk==1.0.24",
    "agntcy-identity-service-sdk==0.0.7",
    "agntcy-app-sdk==0.4.4-dev.1",
    "langfuse>=2.59.0",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=7.0",
    "pytest-asyncio>=0.17.0,<0.18",
    "pytest-cov>=4.0.0,<5",
    "typing-extensions>=4.12.2,<5",
    "openai>=1.86.0,<2.0",
    "autogen-core>=0.4.3,<0.5",
    "coloredlogs>=15.0.1,<16",
    "langchain-openai>=0.3.14,<0.4",
    "sentence-transformers>=5.1.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.poetry]
package-mode = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["farm", "exchange"]
